Align docs asset server configuration
