+++ docs/api/package.json
@@ -30,7 +30,7 @@
     "dev": "NODE_ENV=development pnpm run dev:server",
     "docs": "node src/generate/index.ts",
     "docs:debug": "DEBUG=1 node src/generate/index.ts",
-    "prerender": "node --import remix/node-tsx src/prerender/index.ts",
+    "prerender": "NODE_ENV=production node --import remix/node-tsx src/prerender/index.ts",
     "prerender:serve": "npx http-server -p 44100 build/site",
     "serve": "node --import remix/node-tsx src/server/index.ts",
     "test": "remix test",
